requirements.txt
streamlit
pandas
requests
python-dateutil
altair
numpy
streamlit-autorefresh
kalshi-python
scikit-learn
python-dotenv
joblib

